% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bedloss}
\alias{bedloss}
\title{Mean Squared Error for Genomewide prediction}
\usage{
bedloss(bed, y, weights, pred = NULL, ...)
}
\arguments{
\item{bed}{an rbed object, or a mergedrbed object}

\item{y}{vector of length bed$no.ind describing the truth}

\item{weights}{a vector of length bed$no.snps for an rbed, or length(bed$DATkeep) for a mergedrbed object}

\item{pred}{(default=NULL) a previous call to bedloss, to save calling bedpred again}

\item{...}{extra parameters to \code{\link[pcapred]{get_data}}}
}
\value{
A list containing pred: the predictions and loss: the mse of those predictions
}
\description{
Constructed the genomewide prediction for a score function defined by \code{weights} on an \code{rbed} object, then evaluate  the mean squared error to a provided \code{y}.
}
