% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bedpred}
\alias{bedpred}
\title{Make Genome Wide Predictions From a Bed File}
\usage{
bedpred(bed, weights, ...)
}
\arguments{
\item{bed}{an rbed object, or a mergedrbed object}

\item{weights}{a vector of length bed$no.snps for an rbed, or length(bed$DATkeep) for a mergedrbed object}

\item{...}{extra parameters to \code{\link[pcapred]{get_data}}}
}
\value{
a vector of length bed$no.inds
}
\description{
Create a score for each individual from weighted genetic data
}
