% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{get_loadings}
\alias{get_loadings}
\title{Solve an SVD for loadings}
\usage{
get_loadings(X, U, D)
}
\arguments{
\item{X}{a data matrix of dimension N by L}

\item{U}{the Eigenvector matrix of dimension N by K}

\item{D}{the Eigenvalue vector of length K}
}
\value{
new loadings in the form of an L by K matrix
}
\description{
Take a data matrix \eqn{X}, its corresponding eigenvectors \eqn{U} and eigenvalues \eqn{D}, and return an estimate of \eqn{V}, where \eqn{X= UDV'}.

You might want to do this if: a) you obtained PCA data that did not contain loadings, b) you wish to use a SNP set that is different to the one used to construct the PCA.

This can be dangerous when used for extrapolation, so make sure you test it carefully.
}
