% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{initTheta}
\alias{initTheta}
\title{Generate a random initial parameter vector}
\usage{
initTheta(len, minTheta = -1, maxTheta = 1, seed = NULL)
}
\arguments{
\item{len}{Length of vector}

\item{minTheta}{Minimum of uniform sampling}

\item{maxTheta}{Maximum of uniform sampling}

\item{seed}{(default NULL)}
}
\value{
A matrix with 1 row and len columns
}
\description{
Uniform sampling for initial values of a parameter
}
